add js class for ClassNames object Browser did not see ClassNames object in time to use it in the classnames function. Another way to fix this error was to make it a global.
